Weeks1-4/requirements.txt
alembic==1.18.4
annotated-types==0.7.0
appdirs==1.4.4
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
colorama==0.4.6
colorlog==6.10.1
contourpy==1.3.3
cycler==0.12.1
Cython==3.2.4
docker-pycreds==0.4.0
filelock==3.25.0
filterpy==1.4.5
fonttools==4.61.1
fsspec==2026.2.0
gitdb==4.0.12
GitPython==3.1.46
greenlet==3.3.2
idna==3.11
ImageIO==2.37.3
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.5.0
lap==0.5.13
lazy-loader==0.5
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
narwhals==2.17.0
networkx==3.6.1
numpy==2.4.2
opencv-python==4.13.0.92
optuna==4.7.0
packaging==26.0
pandas==3.0.1
pathtools==0.1.2
pillow==12.1.1
platformdirs==4.9.4
plotly==6.6.0
polars==1.38.1
polars-runtime-32==1.38.1
protobuf==4.25.8
psutil==7.2.2
pyaml==26.2.1
pycocotools==2.0.11
pydantic==2.12.5
pydantic_core==2.41.5
# Editable Git install with no remote (pyflow==1.0)
-e d:\apunts ordinador\master\c6\c6_repo\mcv-c6-2026-team7\external\pyflow
pyparsing==3.3.2
pypng @ git+https://gitlab.com/drj11/pypng@a5844c07da7bebaca833bc984fbe0ed348925666
python-dateutil==2.9.0.post0
PyYAML==6.0.3
requests==2.32.5
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.17.1
sentry-sdk==2.54.0
setproctitle==1.3.7
six==1.17.0
smmap==5.0.3
SQLAlchemy==2.0.48
sympy==1.13.1
tabulate==0.10.0
threadpoolctl==3.6.0
tifffile==2026.3.3
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.65.0
trackeval==1.3.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
ultralytics==8.4.21
ultralytics-thop==2.0.18
urllib3==2.6.3
wandb==0.25.1
yacs==0.1.8
